+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased changes
+
+* Add `earthkit.time.create_sequence` wrapper for easy sequence creation
+
 ## 0.1.8 - 2026-02-16
 
 * Add `earthkit-datetime` tool to compute datetime shifts and differences